cambios de los modelos en sus atribultos de float a double para evitar errores

+++ IDGS903_Api/Models/Gasto_Materia_Prima.cs
@@ -6,8 +6,8 @@ namespace IDGS903_Api.Models
     {
         [Key]
         public int Id { get; set; }
-        public float Cantidad { get; set; }
-        public float Costo { get; set; }
+        public Double Cantidad { get; set; }
+        public Double Costo { get; set; }
         public int MateriaPrima_id { get; set; }
         public int Producto_id { get; set; }
     }
